requirements.txt
Flask
flask-cors
pandas
numpy
tensorflow-cpu
scikit-learn
